+++ lib/src/models/serializable.dart
@@ -11,10 +11,12 @@ abstract class Serializable extends Equatable {
     switch (T) {
       case Video:
         return Video.fromMap(map) as T;
-      case PaginatedResult:
-        return PaginatedResult.fromMap(map) as T;
       case PaginatedResult<VideoWithComments>:
         return PaginatedResult<VideoWithComments>.fromMap(map) as T;
+      case PaginatedResult<VideoFull>:
+        return PaginatedResult<VideoFull>.fromMap(map) as T;
+      case PaginatedResult:
+        return PaginatedResult.fromMap(map) as T;
       case VideoLiveInfo:
         return VideoLiveInfo.fromMap(map) as T;
       case VideoWithComments:
@@ -28,10 +30,12 @@ abstract class Serializable extends Equatable {
     switch (T) {
       case Video:
         return Video.fromJson(source) as T;
-      case PaginatedResult:
-        return PaginatedResult.fromJson(source) as T;
       case PaginatedResult<VideoWithComments>:
         return PaginatedResult<VideoWithComments>.fromJson(source) as T;
+      case PaginatedResult<VideoFull>:
+        return PaginatedResult<VideoFull>.fromJson(source) as T;
+      case PaginatedResult:
+        return PaginatedResult.fromJson(source) as T;
       case VideoLiveInfo:
         return VideoLiveInfo.fromJson(source) as T;
       case VideoWithComments: